budgetReport: Deletes formula evaluation by apache poi
When creating a budget report, the formulas from the template were evaluated by Apache POI, causing formulas to throw an exception without the support of the framework.
The evaluation is not necessary, because Excel evaluates the formulas when opening the file. That's why we deleted the evaluation by Apache POI.