contract: Allow empty values for additional information fields
This also includes a cleanup of the code. Instead of iterating trough
the lists with a nested for-loop, we now create a map with the field
name as a key. This allows us to quickly check if an attribute already
exists, and update it, or create a new one. One disadvantage is, that
the order of the elements isn't guaranteed anymore, but that shouldn't
be a problem.