budget: Change format of aggregation period in budget overview
Before the weekly budget overview only showed the week number
("Week #%d"), with made it hard to distinguish from which month each
record was. Now we prepend the year to the week ("Week %d-%d"). The
format of the monthly overview was changed to be the same
("Month %d/%02d" to "Month %d-%02d").
This also fixes sorting by aggregation period in the monthly overview,
as it now sorts my year and week, instead of just week.