contract: Change ContractBaseData#taxRate from double to BigDecimal
A disadvatange of using the primitive double was that is wasn't possible
to encode a missing value, causing the form to always show 0.0, even if
the value was not set. I decided to use BigDecimal over Double because
the ContractEntity also uses a BigDecimal, which makes mapping easier.
For the divisions the scale has been specified to prevent values from
being truncated when rounding. I chose 4 as the scale because that's
what we use in the database.