person: Fix error when accessing a person details when no records exist
When accessing the details page for a person after all their records
have been deleted a NullPointerException was triggered because the query
in PersonRepository#findDetailDataByPersonId fed the
PersonDetailDataBean constructor null values for valuedMinutes, and
valuedRate, since the table it joins with is empty. As a workaround we
do what is already done in the PersonBaseDataBean constructor. We check
if valuedMinutes is null and, if it is null, default
averageDailyRateInCents to 0.