Remove reflections dependency
This dependency caused warnings when the project was started as a war
file.
Reflections was used to implement the @Mount annotation, to replace
this functionality we now use @MountPath from wicketstuff-annotations.
We now get another warning when starting the project with a war, but
that seems to be an issue with spring and seems has been resolved in
newer versions.