Allow to clear/to have empty dates with DateRangePickerBehavior
It is now possible to to have empty dates and to clear dates with
DateRangePickerBehavior.
To be able to clear dates a new constructor option 'clearable' has been
added. If it is true the cancle button the in the picker gets replaced
with a clear button.
To allow for empty dates the daterangepicker.js library had to be
updated to respect the 'enableEmptyDate' property for non-single
calendars.