ci: Add workflow for publishing nightly releases
This workflow will create a new release every day at 3:24.
This doesn't check if there were any new commits in the past 24 hours,
which we might consider to add as an optimisation in the future.

If you wonder why I chose 3:24, instead of 3:00, or 3:30, that's because
GitHub recommends not running scheduled tasks at the full hour, since
many projects are running theirs during those times, causing workflows
to be delayed. I thought minutes that are a multiple of 5 might suffer
from the same issue, so I decided to use a 'crooked' minute. Initially I
planned to take 3:42, but that was a bit too late for my liking, so
I went with 3:24 instead.